+++ UltitemsCyan/Items/Tier2/XenonAmpoule.cs
@@ -67,7 +67,7 @@ namespace UltitemsCyan.Items.Tier2
                 "XENONAMPOULE",
                 itemName,
                 "Activating your Equipment also fires a laser. Damage and size scale with equipment duration.",
-                "Activating your Equipment also fires a <style=cIsDamage>critting laser</style> for <style=cIsDamage>1600%</style> <style=cStack>(+400% per stack)</style> base damage. The damage scales with an equipment's cooldown.",
+                "Activating your Equipment also fires a <style=cIsDamage>critting laser</style> for <style=cIsDamage>1600%</style> <style=cStack>(+400% per stack)</style> base damage. The damage and size scales with an equipment's cooldown.",
                 "It's Purple because I messed up. Xenon is supposed to be more blue than hyrdogen, but I wanted an X name. Sorry.",
                 ItemTier.Tier2,
                 UltAssets.XenonAmpouleSprite,